Setup initial files for master
This commit is broken. It has an incomplete setup for storybook, but a
functioning setup for Rollup/Typescript. This was pushed on a different
branch called initial-setup, but because I didn't already have a master
branch, git made initial-setup the default branch. The situation has
been corrected, and now updating the commit to reflect something more
descriptivie.