Migrate to Styleguidist from Storybook
This project was initially setup with Storybook as the display for all
of the components. At first this seemed promising, but the fact that
Storybook does not have a live editor nor does it display the API of the
components, it became clear that it was not the right tool for
displaying this library. React Styleguidist provides both of those
things with a built in webpack configruation and scaffold.